code/requirements.txt
mne
mne-bids
openneuro-py
